Split container into multiple modules
